Split i18n files for better modularity and maintainability, restructure translations into separate namespaces, and migrate English and Polish translations into modularized files. Update `types.ts` to reflect the updated structure.
