--- app-v3/src/i18n/en.ts
@@ -1,140 +1,0 @@
-import type { Translations } from './types';
-
-export const en: Translations = {
-  menu: {
-    file: 'File',
-    edit: 'Edit',
-    image: 'Image',
-    layer: 'Layer',
-  },
-  file: {
-    new: 'New Project',
-    open: 'Open Project\u2026',
-    save: 'Save Project',
-    importImage: 'Import Image\u2026',
-    exportPng: 'Export PNG',
-    newConfirm: 'New project? Unsaved changes will be lost.',
-    savedOk: 'Project saved.',
-    openedOk: 'Project opened.',
-    importedOk: 'Image imported.',
-    exportedOk: 'PNG exported.',
-    projectFilename: 'project.wdraft',
-    imageFilename: 'webdraft-image.png',
-  },
-  edit: {
-    undo: 'Undo',
-    redo: 'Redo',
-    cut: 'Cut',
-    copy: 'Copy',
-    paste: 'Paste',
-    clearLayer: 'Clear Layer',
-  },
-  image: {
-    resizeCanvas: 'Resize Canvas\u2026',
-    invertColors: 'Invert Colors',
-    mirrorH: 'Mirror Horizontal',
-    mirrorV: 'Mirror Vertical',
-    rotateLeft: 'Rotate Left',
-    rotateRight: 'Rotate Right',
-    widthPrompt: 'Width (px):',
-    heightPrompt: 'Height (px):',
-    resizedOk: (w, h) => `Canvas: ${w}\u00d7${h}`,
-  },
-  layer: {
-    new: 'New Layer',
-    delete: 'Delete Layer',
-    moveUp: 'Move Layer Up',
-    moveDown: 'Move Layer Down',
-    addedOk: 'Layer added.',
-    deletedOk: 'Layer deleted.',
-  },
-  toolbar: {
-    toolSelect: 'Select \u2014 S',
-    toolPencil: 'Pencil \u2014 P',
-    toolEraser: 'Eraser \u2014 E',
-    toolSampler: 'Sampler \u2014 C',
-    toolFillBucket: 'Fill bucket \u2014 B',
-    toolWeb: 'Web \u2014 W',
-    toolRectangle: 'Rectangle \u2014 R',
-    toolEllipse: 'Ellipse \u2014 O',
-    toolText: 'Text \u2014 T',
-    invertColors: 'Invert colors \u2014 I',
-    rotateLeft: 'Rotate left \u2014 ,',
-    rotateRight: 'Rotate right \u2014 .',
-    mirrorH: 'Mirror horizontally \u2014 H',
-    mirrorV: 'Mirror vertically \u2014 V',
-    size: 'Size',
-    webSensitivity: 'Web',
-    color: 'Color',
-    fillShapes: 'Fill shapes',
-    fillColor: 'Fill color',
-    fillOpacity: 'Fill opacity',
-    bucketTolerance: 'Bucket tolerance',
-    shadow: 'Shadow',
-    shadowColor: 'Shadow color',
-    shadowBlur: 'Shadow blur',
-    shadowX: 'Shadow X',
-    shadowY: 'Shadow Y',
-    textFont: 'Text font',
-    textAlignment: 'Text alignment',
-    bold: 'Bold',
-    italic: 'Italic',
-    clear: 'Clear',
-    undo: 'Undo \u2014 Ctrl+Z',
-    redo: 'Redo \u2014 Ctrl+Shift+Z',
-    copy: 'Copy \u2014 Ctrl+C',
-    cut: 'Cut \u2014 Ctrl+X',
-    paste: 'Paste \u2014 Ctrl+V',
-    width: 'Width',
-    height: 'Height',
-    resizeCanvas: 'Resize canvas',
-    uploadImage: 'Upload image',
-    camera: 'Camera',
-    exportPng: 'Export PNG',
-    saveProject: 'Save project',
-    openProject: 'Open project',
-    pngExportedOk: 'PNG exported.',
-    imageImportedOk: 'Image imported.',
-    canvasResizedOk: (w, h) => `Canvas resized to ${w} \u00d7 ${h}.`,
-  },
-  layers: {
-    title: 'Layers',
-    addLayer: 'Add layer',
-    deleteLayer: 'Delete layer',
-    moveLayerUp: 'Move layer up',
-    moveLayerDown: 'Move layer down',
-    selectLayer: (name) => `Select ${name}`,
-    renameLayer: 'Rename layer',
-    hideLayer: 'Hide layer',
-    showLayer: 'Show layer',
-  },
-  camera: {
-    title: 'Camera',
-    titleWithLabel: (label) => `Camera \u2014 ${label}`,
-    unavailable: 'Camera is unavailable in this browser.',
-    sepia: 'Sepia',
-    noise: 'Noise',
-    greyscale: 'Greyscale',
-    negative: 'Negative',
-    snap: 'Snap',
-    snapToCanvas: 'Snap to canvas',
-    applyToCanvas: 'Apply to canvas',
-    saveToFile: 'Save to file',
-    backToLive: 'Back to live',
-    frameAdded: 'Camera frame added.',
-    cameraError: 'Camera error.',
-    snapFilename: 'camera-snap.png',
-  },
-  picker: {
-    addColor: 'Add current color',
-    restorePrevious: 'Click to restore previous color',
-    newColor: 'New:',
-    previousColor: 'Previous:',
-    range100: '0..100',
-    range255: '0..255',
-  },
-  common: {
-    somethingWentWrong: 'Something went wrong.',
-    appTitle: 'WebDraft',
-  },
-};
