+++ app-v3/src/i18n/en/common.ts
@@ -1,0 +1,6 @@
+import type {CommonTranslations} from '../types.ts';
+
+export const commonTranslations: CommonTranslations = {
+  somethingWentWrong: 'Something went wrong.',
+  appTitle: 'WebDraft',
+};
