+++ app-v3/src/i18n/en/edit.ts
@@ -1,0 +1,10 @@
+import type {EditTranslations} from '../types.ts';
+
+export const editTranslations: EditTranslations = {
+  undo: 'Undo',
+  redo: 'Redo',
+  cut: 'Cut',
+  copy: 'Copy',
+  paste: 'Paste',
+  clearLayer: 'Clear Layer',
+};
