+++ app-v3/src/i18n/en/file.ts
@@ -1,0 +1,16 @@
+import type {FileTranslations} from '../types.ts';
+
+export const fileTranslations: FileTranslations = {
+  new: 'New Project',
+  open: 'Open Project\u2026',
+  save: 'Save Project',
+  importImage: 'Import Image\u2026',
+  exportPng: 'Export PNG',
+  newConfirm: 'New project? Unsaved changes will be lost.',
+  savedOk: 'Project saved.',
+  openedOk: 'Project opened.',
+  importedOk: 'Image imported.',
+  exportedOk: 'PNG exported.',
+  projectFilename: 'project.wdraft',
+  imageFilename: 'webdraft-image.png',
+};
