+++ app-v3/src/i18n/en/layer.ts
@@ -1,0 +1,10 @@
+import type {LayerTranslations} from '../types.ts';
+
+export const layerTranslations: LayerTranslations = {
+  new: 'New Layer',
+  delete: 'Delete Layer',
+  moveUp: 'Move Layer Up',
+  moveDown: 'Move Layer Down',
+  addedOk: 'Layer added.',
+  deletedOk: 'Layer deleted.',
+};
