+++ app-v3/src/i18n/en/layers.ts
@@ -1,0 +1,13 @@
+import type {LayersTranslations} from '../types.ts';
+
+export const layersTranslations: LayersTranslations = {
+  title: 'Layers',
+  addLayer: 'Add layer',
+  deleteLayer: 'Delete layer',
+  moveLayerUp: 'Move layer up',
+  moveLayerDown: 'Move layer down',
+  selectLayer: (name) => `Select ${name}`,
+  renameLayer: 'Rename layer',
+  hideLayer: 'Hide layer',
+  showLayer: 'Show layer',
+};
