+++ app-v3/src/i18n/en/menu.ts
@@ -1,0 +1,8 @@
+import type {MenuTranslations} from '../types.ts';
+
+export const menuTranslations: MenuTranslations = {
+  file: 'File',
+  edit: 'Edit',
+  image: 'Image',
+  layer: 'Layer',
+};
