+++ app-v3/src/i18n/en/picker.ts
@@ -1,0 +1,10 @@
+import type {PickerTranslations} from '../types.ts';
+
+export const pickerTranslations: PickerTranslations = {
+  addColor: 'Add current color',
+  restorePrevious: 'Click to restore previous color',
+  newColor: 'New:',
+  previousColor: 'Previous:',
+  range100: '0..100',
+  range255: '0..255',
+};
