+++ app-v3/src/i18n/en/toolbar.ts
@@ -1,0 +1,51 @@
+import type {ToolbarTranslations} from '../types.ts';
+
+export const toolbarTranslations: ToolbarTranslations = {
+  toolSelect: 'Select \u2014 S',
+  toolPencil: 'Pencil \u2014 P',
+  toolEraser: 'Eraser \u2014 E',
+  toolSampler: 'Sampler \u2014 C',
+  toolFillBucket: 'Fill bucket \u2014 B',
+  toolWeb: 'Web \u2014 W',
+  toolRectangle: 'Rectangle \u2014 R',
+  toolEllipse: 'Ellipse \u2014 O',
+  toolText: 'Text \u2014 T',
+  invertColors: 'Invert colors \u2014 I',
+  rotateLeft: 'Rotate left \u2014 ,',
+  rotateRight: 'Rotate right \u2014 .',
+  mirrorH: 'Mirror horizontally \u2014 H',
+  mirrorV: 'Mirror vertically \u2014 V',
+  size: 'Size',
+  webSensitivity: 'Web',
+  color: 'Color',
+  fillShapes: 'Fill shapes',
+  fillColor: 'Fill color',
+  fillOpacity: 'Fill opacity',
+  bucketTolerance: 'Bucket tolerance',
+  shadow: 'Shadow',
+  shadowColor: 'Shadow color',
+  shadowBlur: 'Shadow blur',
+  shadowX: 'Shadow X',
+  shadowY: 'Shadow Y',
+  textFont: 'Text font',
+  textAlignment: 'Text alignment',
+  bold: 'Bold',
+  italic: 'Italic',
+  clear: 'Clear',
+  undo: 'Undo \u2014 Ctrl+Z',
+  redo: 'Redo \u2014 Ctrl+Shift+Z',
+  copy: 'Copy \u2014 Ctrl+C',
+  cut: 'Cut \u2014 Ctrl+X',
+  paste: 'Paste \u2014 Ctrl+V',
+  width: 'Width',
+  height: 'Height',
+  resizeCanvas: 'Resize canvas',
+  uploadImage: 'Upload image',
+  camera: 'Camera',
+  exportPng: 'Export PNG',
+  saveProject: 'Save project',
+  openProject: 'Open project',
+  pngExportedOk: 'PNG exported.',
+  imageImportedOk: 'Image imported.',
+  canvasResizedOk: (w, h) => `Canvas resized to ${w} \u00d7 ${h}.`,
+};
