--- app-v3/src/i18n/pl.ts
@@ -1,140 +1,0 @@
-import type { Translations } from './types';
-
-export const pl: Translations = {
-  menu: {
-    file: 'Plik',
-    edit: 'Edycja',
-    image: 'Obraz',
-    layer: 'Warstwa',
-  },
-  file: {
-    new: 'Nowy projekt',
-    open: 'Otwórz projekt\u2026',
-    save: 'Zapisz projekt',
-    importImage: 'Importuj obraz\u2026',
-    exportPng: 'Eksportuj PNG',
-    newConfirm: 'Nowy projekt? Niezapisane zmiany zostaną utracone.',
-    savedOk: 'Projekt zapisany.',
-    openedOk: 'Projekt otwarty.',
-    importedOk: 'Obraz zaimportowany.',
-    exportedOk: 'PNG wyeksportowany.',
-    projectFilename: 'projekt.wdraft',
-    imageFilename: 'webdraft-obraz.png',
-  },
-  edit: {
-    undo: 'Cofnij',
-    redo: 'Ponów',
-    cut: 'Wytnij',
-    copy: 'Kopiuj',
-    paste: 'Wklej',
-    clearLayer: 'Wyczyść warstwę',
-  },
-  image: {
-    resizeCanvas: 'Zmień rozmiar płótna\u2026',
-    invertColors: 'Odwróć kolory',
-    mirrorH: 'Lustro poziome',
-    mirrorV: 'Lustro pionowe',
-    rotateLeft: 'Obróć w lewo',
-    rotateRight: 'Obróć w prawo',
-    widthPrompt: 'Szerokość (px):',
-    heightPrompt: 'Wysokość (px):',
-    resizedOk: (w, h) => `Płótno: ${w}\u00d7${h}`,
-  },
-  layer: {
-    new: 'Nowa warstwa',
-    delete: 'Usuń warstwę',
-    moveUp: 'Przesuń warstwę w górę',
-    moveDown: 'Przesuń warstwę w dół',
-    addedOk: 'Warstwa dodana.',
-    deletedOk: 'Warstwa usunięta.',
-  },
-  toolbar: {
-    toolSelect: 'Zaznaczanie — S',
-    toolPencil: 'Ołówek — P',
-    toolEraser: 'Gumka — E',
-    toolSampler: 'Próbnik — C',
-    toolFillBucket: 'Wiadro z farbą — B',
-    toolWeb: 'Sieć — W',
-    toolRectangle: 'Prostokąt — R',
-    toolEllipse: 'Elipsa — O',
-    toolText: 'Tekst — T',
-    invertColors: 'Odwróć kolory — I',
-    rotateLeft: 'Obróć w lewo — ,',
-    rotateRight: 'Obróć w prawo — .',
-    mirrorH: 'Lustro poziome — H',
-    mirrorV: 'Lustro pionowe — V',
-    size: 'Rozmiar',
-    webSensitivity: 'Sieć',
-    color: 'Kolor',
-    fillShapes: 'Wypełnij kształty',
-    fillColor: 'Kolor wypełnienia',
-    fillOpacity: 'Krycie wypełnienia',
-    bucketTolerance: 'Tolerancja',
-    shadow: 'Cień',
-    shadowColor: 'Kolor cienia',
-    shadowBlur: 'Rozmycie cienia',
-    shadowX: 'Cień X',
-    shadowY: 'Cień Y',
-    textFont: 'Czcionka',
-    textAlignment: 'Wyrównanie',
-    bold: 'Pogrubienie',
-    italic: 'Kursywa',
-    clear: 'Wyczyść',
-    undo: 'Cofnij — Ctrl+Z',
-    redo: 'Ponów — Ctrl+Shift+Z',
-    copy: 'Kopiuj — Ctrl+C',
-    cut: 'Wytnij — Ctrl+X',
-    paste: 'Wklej — Ctrl+V',
-    width: 'Szerokość',
-    height: 'Wysokość',
-    resizeCanvas: 'Zmień rozmiar płótna',
-    uploadImage: 'Wgraj obraz',
-    camera: 'Kamera',
-    exportPng: 'Eksportuj PNG',
-    saveProject: 'Zapisz projekt',
-    openProject: 'Otwórz projekt',
-    pngExportedOk: 'PNG wyeksportowany.',
-    imageImportedOk: 'Obraz zaimportowany.',
-    canvasResizedOk: (w, h) => `Płótno: ${w}\u00d7${h}`,
-  },
-  layers: {
-    title: 'Warstwy',
-    addLayer: 'Dodaj warstwę',
-    deleteLayer: 'Usuń warstwę',
-    moveLayerUp: 'Przesuń warstwę w górę',
-    moveLayerDown: 'Przesuń warstwę w dół',
-    selectLayer: (name) => `Zaznacz ${name}`,
-    renameLayer: 'Zmień nazwę warstwy',
-    hideLayer: 'Ukryj warstwę',
-    showLayer: 'Pokaż warstwę',
-  },
-  camera: {
-    title: 'Kamera',
-    titleWithLabel: (label) => `Kamera \u2014 ${label}`,
-    unavailable: 'Kamera jest niedostępna w tej przeglądarce.',
-    sepia: 'Sepia',
-    noise: 'Szum',
-    greyscale: 'Skala szarości',
-    negative: 'Negatyw',
-    snap: 'Zdjęcie',
-    snapToCanvas: 'Zdjęcie na płótno',
-    applyToCanvas: 'Zastosuj na płótno',
-    saveToFile: 'Zapisz do pliku',
-    backToLive: 'Wróć do podglądu',
-    frameAdded: 'Kadr z kamery dodany.',
-    cameraError: 'Błąd kamery.',
-    snapFilename: 'camera-snap.png',
-  },
-  picker: {
-    addColor: 'Dodaj bieżący kolor',
-    restorePrevious: 'Kliknij, aby przywrócić poprzedni kolor',
-    newColor: 'Nowy:',
-    previousColor: 'Poprzedni:',
-    range100: '0..100',
-    range255: '0..255',
-  },
-  common: {
-    somethingWentWrong: 'Coś poszło nie tak.',
-    appTitle: 'WebDraft',
-  },
-};
