+++ app-v3/src/i18n/pl/common.ts
@@ -1,0 +1,6 @@
+import type {CommonTranslations} from '../types.ts';
+
+export const commonTranslations: CommonTranslations = {
+  somethingWentWrong: 'Coś poszło nie tak.',
+  appTitle: 'WebDraft',
+};
