+++ app-v3/src/i18n/pl/edit.ts
@@ -1,0 +1,10 @@
+import type {EditTranslations} from '../types.ts';
+
+export const editTranslations: EditTranslations = {
+  undo: 'Cofnij',
+  redo: 'Ponów',
+  cut: 'Wytnij',
+  copy: 'Kopiuj',
+  paste: 'Wklej',
+  clearLayer: 'Wyczyść warstwę',
+};
