+++ app-v3/src/i18n/pl/file.ts
@@ -1,0 +1,16 @@
+import type {FileTranslations} from '../types.ts';
+
+export const fileTranslations: FileTranslations = {
+  new: 'Nowy projekt',
+  open: 'Otwórz projekt\u2026',
+  save: 'Zapisz projekt',
+  importImage: 'Importuj obraz\u2026',
+  exportPng: 'Eksportuj PNG',
+  newConfirm: 'Nowy projekt? Niezapisane zmiany zostaną utracone.',
+  savedOk: 'Projekt zapisany.',
+  openedOk: 'Projekt otwarty.',
+  importedOk: 'Obraz zaimportowany.',
+  exportedOk: 'PNG wyeksportowany.',
+  projectFilename: 'projekt.wdraft',
+  imageFilename: 'webdraft-obraz.png',
+};
