+++ app-v3/src/i18n/pl/layer.ts
@@ -1,0 +1,10 @@
+import type {LayerTranslations} from '../types.ts';
+
+export const layerTranslations: LayerTranslations = {
+  new: 'Nowa warstwa',
+  delete: 'Usuń warstwę',
+  moveUp: 'Przesuń warstwę w górę',
+  moveDown: 'Przesuń warstwę w dół',
+  addedOk: 'Warstwa dodana.',
+  deletedOk: 'Warstwa usunięta.',
+};
