+++ app-v3/src/i18n/pl/menu.ts
@@ -1,0 +1,8 @@
+import type {MenuTranslations} from '../types.ts';
+
+export const menuTranslations: MenuTranslations = {
+  file: 'Plik',
+  edit: 'Edycja',
+  image: 'Obraz',
+  layer: 'Warstwa',
+};
