+++ app-v3/src/i18n/pl/picker.ts
@@ -1,0 +1,10 @@
+import type {PickerTranslations} from '../types.ts';
+
+export const pickerTranslations: PickerTranslations = {
+  addColor: 'Dodaj bieżący kolor',
+  restorePrevious: 'Kliknij, aby przywrócić poprzedni kolor',
+  newColor: 'Nowy:',
+  previousColor: 'Poprzedni:',
+  range100: '0..100',
+  range255: '0..255',
+};
