+++ app-v3/src/i18n/pl/toolbar.ts
@@ -1,0 +1,51 @@
+import type {ToolbarTranslations} from '../types.ts';
+
+export const toolbarTranslations: ToolbarTranslations = {
+  toolSelect: 'Zaznaczanie \u2014 S',
+  toolPencil: 'Ołówek \u2014 P',
+  toolEraser: 'Gumka \u2014 E',
+  toolSampler: 'Próbnik \u2014 C',
+  toolFillBucket: 'Wiadro z farbą \u2014 B',
+  toolWeb: 'Sieć \u2014 W',
+  toolRectangle: 'Prostokąt \u2014 R',
+  toolEllipse: 'Elipsa \u2014 O',
+  toolText: 'Tekst \u2014 T',
+  invertColors: 'Odwróć kolory \u2014 I',
+  rotateLeft: 'Obróć w lewo \u2014 ,',
+  rotateRight: 'Obróć w prawo \u2014 .',
+  mirrorH: 'Lustro poziome \u2014 H',
+  mirrorV: 'Lustro pionowe \u2014 V',
+  size: 'Rozmiar',
+  webSensitivity: 'Sieć',
+  color: 'Kolor',
+  fillShapes: 'Wypełnij kształty',
+  fillColor: 'Kolor wypełnienia',
+  fillOpacity: 'Krycie wypełnienia',
+  bucketTolerance: 'Tolerancja',
+  shadow: 'Cień',
+  shadowColor: 'Kolor cienia',
+  shadowBlur: 'Rozmycie cienia',
+  shadowX: 'Cień X',
+  shadowY: 'Cień Y',
+  textFont: 'Czcionka',
+  textAlignment: 'Wyrównanie',
+  bold: 'Pogrubienie',
+  italic: 'Kursywa',
+  clear: 'Wyczyść',
+  undo: 'Cofnij \u2014 Ctrl+Z',
+  redo: 'Ponów \u2014 Ctrl+Shift+Z',
+  copy: 'Kopiuj \u2014 Ctrl+C',
+  cut: 'Wytnij \u2014 Ctrl+X',
+  paste: 'Wklej \u2014 Ctrl+V',
+  width: 'Szerokość',
+  height: 'Wysokość',
+  resizeCanvas: 'Zmień rozmiar płótna',
+  uploadImage: 'Wgraj obraz',
+  camera: 'Kamera',
+  exportPng: 'Eksportuj PNG',
+  saveProject: 'Zapisz projekt',
+  openProject: 'Otwórz projekt',
+  pngExportedOk: 'PNG wyeksportowany.',
+  imageImportedOk: 'Obraz zaimportowany.',
+  canvasResizedOk: (w, h) => `Płótno zmienione: ${w}\u00d7${h}`,
+};
